second Changes for 1.5
added Gostana (Kingom of Khotan) and removed other releasable in the same provinces. And made a special horse (horse because of the coins found there) flag, tried to use colors used in symbols in paintings etc. found there.

changed Interventionism to Traditionalism for Iran. (its historical) and will come to vanilla in 1.5 aswell.

Bank-e_Melli-e_Iran as Company

Made some major Changes to Provinces in Iran (mines agriculture etc.), and  changed the files to match new 1.5 files for Vineyards etc.