Localizations, name changes, set custom tooltips
change trait "Zoroastrian" to "Iranic" for religion, for the future to add Kafkasian and Turan Iranic of origin religions.
set custom tooltips for decisions variables.
localization of landowners.
localization of nomads event,  journal entry, new decisions and modifiers.

last commit also included a new image for the thumbnail, created an image of Shah Square in front of Damavand. Some fire/light inside the Shah Square to represent the Mazda Yasna.